moved CampyContext to the src folder
