+++ src/App.js
@@ -20,7 +20,7 @@ import {UploadLocationPics} from'./components/protectedRoutes/UploadLocationPics
 import { Footer } from "./components/Footer";
 import { ProfilePicUpload } from "./components/protectedRoutes/ProfilePicUpload";
 import { ProtectedRoute } from "./components/auth/ProtectedRoute";
-import { CampyContext } from "./context/CampyContext";
+import { CampyContext } from "./CampyContext";
 import { CssBaseline } from "@material-ui/core";
 import { ThemeProvider } from "@material-ui/core/styles";
 import theme from "./theme";
